+++ apps/web/app/(app)/[chain]/[garden]/[community]/[poolId]/[proposalId]/page.tsx
