+++ apps/web/hooks/useChainIdFromPath.ts
@@ -2,6 +2,6 @@ import { usePathname } from "next/navigation";
 
 export function useChainIdFromPath() {
   const path = usePathname();
-  const segment = path?.split("/")[2];
+  const segment = path?.split("/")[1];
   return segment ? Number(segment) : undefined;
 }
